int16 to int8

+++ parkir v1/program01/src/main.cpp
@@ -14,7 +14,7 @@ const int ledQ = 12;
 
 Event Send = Event(500);
 
-int pulse;
+uint8_t pulse;
 unsigned long previousMillis;
 
 void setup()
